+++ libtonode-tests/tests/concrete.rs
@@ -618,7 +618,8 @@ async fn test_scanning_in_watch_only_mode() {
                 birthday: 1,
                 wallet_settings: default_test_wallet_settings(),
             })
-            .build();
+            .build()
+            .unwrap();
         let mut watch_client = LightClient::new(zingo_config, false).await.unwrap();
         // assert empty wallet before rescan
         let balance = watch_client
@@ -1618,7 +1619,8 @@ mod testnet_test {
                     wallet_settings: default_test_wallet_settings(),
                 })
                 .set_wallet_dir(wallet_dir.path().to_path_buf())
-                .build();
+                .build()
+                .unwrap();
 
             let mut lightclient = LightClient::new(config, true).await.unwrap();
             lightclient.save_task().await;
@@ -1644,7 +1646,8 @@ mod testnet_test {
                 .set_indexer_uri((DEFAULT_INDEXER_URI_TESTNET).parse::<http::Uri>().unwrap())
                 .set_wallet_config(WalletConfig::Read)
                 .set_wallet_dir(wallet_dir.path().to_path_buf())
-                .build();
+                .build()
+                .unwrap();
             LightClient::new(config, true).await.unwrap();
 
             test_count += 1;